diff: move materialized_diff_stream() to jj_lib::conflicts module
New diff_contains() revset function will use this helper.
